Move address types to shared
Related to #7 and #134
