+++ .changeset/famous-mayflies-care.md
@@ -1,0 +1,5 @@
+---
+"@alephium/shared": patch
+---
+
+Move address types from all apps
