+++ apps/desktop-wallet/src/hooks/useAddressGeneration.tsx
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AddressKeyPair, deriveNewAddressData, getWalletFromMnemonic } from '@alephium/shared'
+import { AddressKeyPair, AddressMetadata, deriveNewAddressData, getWalletFromMnemonic } from '@alephium/shared'
 import { TOTAL_NUMBER_OF_GROUPS } from '@alephium/web3'
 
 import { useAppDispatch, useAppSelector } from '@/hooks/redux'
@@ -30,7 +30,7 @@ import { selectAllAddresses } from '@/storage/addresses/addressesSelectors'
 import { saveNewAddresses } from '@/storage/addresses/addressesStorageUtils'
 import AddressMetadataStorage from '@/storage/addresses/addressMetadataPersistentStorage'
 import { getEncryptedStoragePropsFromActiveWallet } from '@/storage/encryptedPersistentStorage'
-import { AddressBase, AddressMetadata } from '@/types/addresses'
+import { AddressBase } from '@/types/addresses'
 import { getInitialAddressSettings } from '@/utils/addresses'
 import { getRandomLabelColor } from '@/utils/colors'
