+++ apps/desktop-wallet/src/storage/addresses/addressMetadataPersistentStorage.ts
@@ -16,11 +16,12 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
+import { AddressMetadata, AddressSettings } from '@alephium/shared'
+
 import {
   getEncryptedStoragePropsFromActiveWallet,
   PersistentEncryptedStorage
 } from '@/storage/encryptedPersistentStorage'
-import { AddressMetadata, AddressSettings } from '@/types/addresses'
 import { latestAddressMetadataVersion } from '@/utils/migration'
 
 interface AddressMetadataStorageStoreProps {
