+++ apps/desktop-wallet/src/storage/addresses/addressesActions.ts
@@ -16,7 +16,14 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ADDRESSES_QUERY_LIMIT, AddressHash, CHART_DATE_FORMAT, getHumanReadableError } from '@alephium/shared'
+import {
+  ADDRESSES_QUERY_LIMIT,
+  AddressHash,
+  AddressSettings,
+  BalanceHistory,
+  CHART_DATE_FORMAT,
+  getHumanReadableError
+} from '@alephium/shared'
 import { explorer } from '@alephium/web3'
 import { createAction, createAsyncThunk } from '@reduxjs/toolkit'
 import dayjs from 'dayjs'
@@ -34,14 +41,7 @@ import client from '@/api/client'
 import i18n from '@/i18n'
 import { selectAddressByHash, selectAllAddresses } from '@/storage/addresses/addressesSelectors'
 import { RootState } from '@/storage/store'
-import {
-  Address,
-  AddressBase,
-  AddressSettings,
-  AddressTransactionsSyncResult,
-  BalanceHistory,
-  LoadingEnabled
-} from '@/types/addresses'
+import { Address, AddressBase, AddressTransactionsSyncResult, LoadingEnabled } from '@/types/addresses'
 import { Contact } from '@/types/contacts'
 import { Message, SnackbarMessage } from '@/types/snackbar'
