+++ apps/mobile-wallet/src/store/addresses/addressesStorageUtils.ts
@@ -16,11 +16,10 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { deriveNewAddressData, walletImportAsyncUnsafe } from '@alephium/shared'
+import { AddressMetadata, deriveNewAddressData, walletImportAsyncUnsafe } from '@alephium/shared'
 
 import { newAddressGenerated, syncAddressesData, syncAddressesHistoricBalances } from '~/store/addressesSlice'
 import { store } from '~/store/store'
-import { AddressMetadata } from '~/types/addresses'
 import { Mnemonic, WalletMetadata } from '~/types/wallet'
 import { persistAddressesSettings } from '~/utils/addresses'
 import { mnemonicToSeed } from '~/utils/crypto'
