+++ apps/mobile-wallet/src/types/wallet.ts
@@ -16,9 +16,9 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AddressKeyPair } from '@alephium/shared'
+import { AddressKeyPair, AddressMetadata } from '@alephium/shared'
 
-import { AddressMetadata, AddressPartial } from '~/types/addresses'
+import { AddressPartial } from '~/types/addresses'
 import { Contact, ContactFormData } from '~/types/contacts'
 
 export type Mnemonic = string
