+++ packages/shared/lib/address.ts
@@ -17,11 +17,38 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { bs58, ExplorerProvider, TOTAL_NUMBER_OF_GROUPS } from '@alephium/web3'
+import { AddressInfo, AddressTokenBalance } from '@alephium/web3/dist/src/api/api-explorer'
 import { BIP32Interface } from 'bip32'
 
 import { AddressKeyPair, deriveNewAddressData } from './wallet'
 
 export type AddressHash = string
+
+export type AddressSettings = {
+  isDefault: boolean
+  color: string
+  label?: string
+}
+
+export type AddressIndex = number
+
+export type AddressMetadata = AddressSettings & {
+  index: AddressIndex
+}
+
+export type AddressBalancesSyncResult = Omit<AddressInfo, 'txNumber'> & {
+  hash: AddressHash
+}
+
+export type AddressTokensSyncResult = {
+  hash: AddressHash
+  tokenBalances: AddressTokenBalance[]
+}
+
+export type BalanceHistory = {
+  date: string // CHART_DATE_FORMAT
+  balance: string
+}
 
 export const isAddressValid = (address: string) =>
   !!address && /^[1-9A-HJ-NP-Za-km-z]+$/.test(address) && bs58.decode(address).slice(1).length >= 32
